Day 17 final - based on https://github.com/bradwilson/advent-2022

+++ AOC.Common/P2.cs
@@ -1,6 +1,6 @@
 namespace AOC.Common;
 
-public record struct P2(int X, int Y)
+public record struct P2(long X, long Y)
 {
     public static P2 operator +(P2 p1, P2 p2) =>
         new(p1.X + p2.X, p1.Y + p2.Y);
@@ -11,8 +11,8 @@ public record struct P2(int X, int Y)
     public static P2 operator *(P2 vector, int scalar) =>
         new(vector.X * scalar, vector.Y * scalar);
 
-    public static int Manhattan(P2 p1, P2 p2) =>
-        int.Abs(p1.X - p2.X) + int.Abs(p1.Y - p2.Y);
+    public static long Manhattan(P2 p1, P2 p2) =>
+        long.Abs(p1.X - p2.X) + long.Abs(p1.Y - p2.Y);
 
     public static IEnumerable<P2> ReverseManhattan(P2 point, int distance) =>
         Enumerable.Range(0, distance).SelectMany(i => new[]
